# Conflicts:
#   Ressources.qrc
#   filterarea.cpp
#   imagedit.cpp
#   imagedit.h
#   imagedit.ui

+++ imagedit.cpp
@@ -5,17 +5,21 @@
 ImagEdit::ImagEdit(QWidget *parent) : QMainWindow(parent), ui(new Ui::ImagEdit)
 {
     ui->setupUi(this);
+    ui->tabWidget->setStyleSheet("background-color:black;");
     path = new QString();
     pix = new QPixmap(*path);
     QList<QPushButton*> tousLesBoutons = findChildren<QPushButton*>();
+    confirmCropping = new QPushButton("Rogner", this);
+    confirmCropping->hide();
+    cancelCropping = new QPushButton("Annuler", this);
+    cancelCropping->hide();
 
-    for (QPushButton* bouton : tousLesBoutons) {
-        bouton->setStyleSheet("QPushButton:pressed {border: 2px solid #00f;}"
-                              "QPushButton { border: 1px solid #000;}");
-    }
-
+    boutonResize = findChild<QPushButton*>("resize");
+    resize = new Resize(this);
+    setupResizeButtonConnection();
     raccourciEnregistrer = new QShortcut(QKeySequence(Qt::CTRL | Qt::Key_S), this);
     raccourciOuvrir = new QShortcut(QKeySequence(Qt::CTRL | Qt::Key_O), this);
+    racourciEchapCropping = new QShortcut(QKeySequence(Qt::Key_Escape), this);
 
     connect(raccourciEnregistrer, &QShortcut::activated, this, &ImagEdit::on_save_clicked);
     connect(raccourciOuvrir, &QShortcut::activated, this, &ImagEdit::on_open_clicked);
@@ -37,24 +41,61 @@ ImagEdit::ImagEdit(QWidget *parent) : QMainWindow(parent), ui(new Ui::ImagEdit)
     boutonFiltre = findChild<QPushButton*>("filter");
     widgetFilter = new FilterArea(this);
     setupFilterButtonConnection();
+
+    boutonResize = findChild<QPushButton*>("resize");
+    resize = new Resize(this);
+    setupResizeButtonConnection();
+    resize->hide();
+
 }
 
 
+
+void ImagEdit::setupResizeButtonConnection()
+{
+    resize->setVisible(false);
+    int x = 770;
+    int y = 130;
+    resize->move(x, y);
+
+
+}
 
 void ImagEdit::on_open_clicked()
 {
     /*
     QString cheminInitial = QStandardPaths::writableLocation(QStandardPaths::PicturesLocation);
-    QString cheminFichier = QFileDialog::getOpenFileName(this, "Sélectionnez un fichier", cheminInitial);
+
+    QString selfilter = tr("JPEG (*.jpg *.jpeg)");
+    QString cheminFichier = QFileDialog::getOpenFileName(
+        this,
+        "Sélectionnez un fichier",
+        cheminInitial,
+        tr("All files (*.*);;JPEG (*.jpg *.jpeg);;TIFF (*.tif)" ),
+        &selfilter
+        );
+
     QFileInfo fileInfo(cheminFichier);
     *fileName = fileInfo.fileName();
-    *path = cheminFichier;
-    originalPaths.append(*path);
-    QPushButton *button = new QPushButton();
 
-    croppingButtons.append(button);
+    if(!fileName->isEmpty()) {
+        ui->importImage->close();
+        *path = cheminFichier;
+        originalPaths.append(*path);
+        pix = new QPixmap(*path);
+        *pix = pix->scaled(gridSize, gridSize, Qt::KeepAspectRatio);
 
-    const int gridSize = 30;
+        QPushButton *button = new QPushButton();
+        croppingButtons.append(button);
+
+        button->setIcon(QIcon(*pix));
+        button->setIconSize(pix->size());
+        button->setStyleSheet("QPushButton:pressed {border: 1px solid #00f;} ""QPushButton { border: 1px solid ;}");
+        button->setFixedSize(50, 50);
+        ui->gridLayout->addWidget(button);
+
+        if(ui->gridLayout->count() == 1) displayOnEdition();
+        else updateLibraryVisualisation();
 
     pix = new QPixmap(*path);
     *pix = pix->scaled(gridSize, gridSize, Qt::KeepAspectRatio);
@@ -103,6 +144,7 @@ void ImagEdit::on_open_clicked()
 
 void ImagEdit::updateLibraryVisualisation() {
     int c = 0;
+    qDebug() << croppingButtons.size();
     for(int i = 0; i < ui->gridLayout->count(); i++) {
         if(i % 3 == 0 && i != 0) c++;
 
@@ -263,26 +305,47 @@ void ImagEdit::on_save_clicked()
 void ImagEdit::setupFilterButtonConnection()
 {
     widgetFilter->setVisible(false);
-    int x = 750;
-    int y = 100;
+    int x = 770;
+    int y = 130;
     widgetFilter->move(x, y);
 }
 
 void ImagEdit::on_filter_clicked()
 {
-    widgetFilter->setLabel(actualCropping);
-    widgetFilter->show();
+    if(ui->tabWidget->count() == 0) {
+        QMessageBox::warning(this, "No image found", "No image selected");
+    } else {
+        widgetFilter->setIsFilter(true);
+        widgetFilter->setLabel(actualCropping);
+        widgetFilter->show();
+    }
 }
 
+
 void ImagEdit::on_rogner_clicked() {
-    if(!actualCropping->getIsCropping()) {
-        actualCropping->drawRectCropping(pix);
+    if(ui->tabWidget->count() == 0) {
+        QMessageBox::warning(this, "No image found", "No image selected");
+    } else {
+        if(!actualCropping->getIsCropping()) {
+            confirmCropping->move(640, 180);
+            confirmCropping->show();
+            cancelCropping->move(520, 180);
+            cancelCropping->show();
+            connect(confirmCropping, SIGNAL(clicked()), actualCropping, SLOT(resizePicture()));
+            connect(cancelCropping, SIGNAL(clicked()), this, SLOT(on_rogner_clicked()));
+            connect(racourciEchapCropping, &QShortcut::activated, this, &ImagEdit::on_rogner_clicked);
+            actualCropping->drawRectCropping(pix);
+        }
+        else {
+            confirmCropping->hide();
+            cancelCropping->hide();
+            disconnect(racourciEchapCropping, &QShortcut::activated, this, &ImagEdit::on_rogner_clicked);
+            actualCropping->deleteRectCropping();
+            *pix = actualCropping->getPixmap();
+            update();
+        }
     }
-    else {
-        actualCropping->deleteRectCropping();
-        *pix = actualCropping->getPixmap();
-        update();
-    }
+
 }
 
 void ImagEdit::updateIndicesAfterRemoval(int removedIndex)
